Update decorators to wrap docstrings and reorganise into their own module
